feat(p10,10Z): widen ZirInstr.payload u32 → u64 per ROADMAP §10 row 10.Z
Cycle-2 of the architectural-chunk attempt (per
`.dev/phase10_z_chunk_plan.md` §"Cycle-2 strategy"; cycle-1
measured the cascade + reverted at `83f030ee`). Subagent-driven
mechanical migration completed the widen.

ZirInstr.payload (`src/ir/zir.zig:73`) widened from `u32` to
`u64`. The cycle-1 cascade of 131 compile errors closed with the
following pattern:

  - **Helper signature widen** (memory load/store + bulk_memory):
    `loadInt(rt, W, sign_extend, offset: u64)` and
    `effectiveAddrStore(offset: u64, ...)` widened their offset
    param. Downstream `rt.memory[offset..]` access takes `usize`
    (already u64 on Mac aarch64 / Linux x86_64 / Windows x86_64).
  - **Codegen narrow at consumer**: arm64 + x86_64 emit sites
    that pack the payload into a u32 field (funcidx / typeidx /
    elemidx / dataidx / tableidx / local_idx / lane_idx) cast
    via `@intCast(ins.payload)` at the variable declaration.
  - **i32.const lowering**: `@truncate(ins.payload)` narrows
    the low 32 bits for MOV-imm32 emit on x86_64 + the arm64
    `emitConstU32` site. High 32 bits are zero for any
    Wasm 1.0/2.0 source (i32.const encoded as 32-bit signed LEB).
  - **`@bitCast` size mismatch resolves**: payload writes from
    parser (`mvp/mod.zig::parseI32Const` + `parseI64Const`)
    add an explicit `@as(u32, ...)` cast before assignment so
    high bits stay zero per the existing low-32 / extra hi-32
    encoding contract. Test fixtures (dispatch / mvp / interp
    sites) wrap their `@as(i32, K)` source through the same
    explicit narrow.
  - **LowerInstr.payload mirror** (`zir.zig:225`) widened.

emit_test_*.zig byte-identical maintained — regalloc + emit
only consume the low 32 bits of payload for Wasm 1.0/2.0
sources; the JIT-emitted machine code stays identical for
every fixture under `src/engine/codegen/{arm64,x86_64}/
emit_test*.zig`. Verified by `zig build test-all` GREEN
(1773/1787 PASS, 14 skipped) on Mac aarch64.

30 files modified:
  src/ir/{zir,lower,analysis/{const_prop,liveness}}.zig
  src/interp/{dispatch,mvp}.zig
  src/feature/mvp/mod.zig
  src/instruction/wasm_1_0/{memory,numeric_int}.zig
  src/engine/codegen/arm64/{emit,op_call,op_const,op_globals,
    op_memory,op_simd,op_simd_float,op_simd_int_cmp_lane,op_table}.zig
  src/engine/codegen/shared/{regalloc_shape_tags,regalloc_vreg_class}.zig
  src/engine/codegen/x86_64/{emit,emit_test_float,op_alu_float,
    op_alu_int,op_call,op_control,op_globals,op_locals,op_memory,
    op_table}.zig

Logic-class chunk per classify_chunk_scope.sh. Mac `test-all`
foreground GREEN; ubuntu kicked post-push (test-all).

Mac `zig build test`: 1827/1841 PASS (substrate baseline).
Mac `zig build test-all`: 1773/1787 PASS, 14 skipped, 42/42
build steps.
Mac `zig build lint -- --max-warnings 0`: clean.
Phase 9 invariants: 18/18 PASS.

`ZirInstr.payload` now carries memory64 offsets at spec-full
width per ROADMAP §10 / 10.Z + the Phase 10 design plan §3.1.
10.Z attempt 2/3 SUCCESSFUL — `[x]` flip lands at the row's
close commit (next handover cycle).

Refs: ROADMAP §10 / 10.Z, `phase10_z_chunk_plan.md`,
`phase10_design_plan_ja.md` §3.1, `.claude/rules/architectural_spike.md`.

+++ src/ir/lower.zig
@@ -485,7 +485,7 @@ pub const Lowerer = struct {
     }
 
     // SIBLING-PUB: lower_simd.zig (per ADR-0089 extraction)
-    pub fn emit(self: *Lowerer, op: ZirOp, payload: u32, extra: u32) Error!void {
+    pub fn emit(self: *Lowerer, op: ZirOp, payload: u64, extra: u32) Error!void {
         // D-093 (d-1): skip ZirInstr emission while in the dead
         // region following an unconditional terminator. Operand
         // bytes are still consumed by the caller (the dispatch
